% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_input_dependencies.R
\name{extract_input_dependencies}
\alias{extract_input_dependencies}
\title{Extract input dependencies from a Shiny render block}
\usage{
extract_input_dependencies(render_block)
}
\arguments{
\item{render_block}{A character vector representing a block of Shiny server code.}
}
\value{
A string of unique input IDs used within the render block, or \code{"None"} if no inputs are found.
}
\description{
Identifies all \verb{input$} dependencies within a render block, including cases like \code{input$var} and \code{input[['var']]}.
}
