% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_named_reactives.R
\name{extract_named_reactives}
\alias{extract_named_reactives}
\title{Extract named reactive expressions from a Shiny server file}
\usage{
extract_named_reactives(server_code)
}
\arguments{
\item{server_code}{A character vector representing lines of Shiny server code.}
}
\value{
A data frame containing named reactive expressions with their corresponding input dependencies.
}
\description{
Identifies \code{reactive()} assignments that are stored in variables and returns a data frame.
Additionally, it checks for any input dependencies inside the \code{reactive()} function.
}
