% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output_assignments.R
\name{extract_output_assignments}
\alias{extract_output_assignments}
\title{Extract output assignments from Shiny server code}
\usage{
extract_output_assignments(server_code)
}
\arguments{
\item{server_code}{A character vector representing lines of Shiny server code.}
}
\value{
A list of lists, where each list contains an index, output ID, and render function type.
}
\description{
Identifies \verb{output$} assignments in a Shiny server file and extracts
their associated render function types.
}
