% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_extract_ui_features.R
\name{extract_ui_features}
\alias{extract_ui_features}
\title{Extract Shiny UI input and output elements and their ids (including custom Output functions)}
\usage{
extract_ui_features(ui_code)
}
\arguments{
\item{ui_code}{A character vector representing lines of Shiny UI code.}
}
\value{
A data frame containing the input/output functions and their associated ids.
}
\description{
This function identifies all Shiny input functions (e.g., selectInput, textInput, numericInput) and output functions
(e.g., plotOutput, tableOutput, or custom output functions) in a UI script, and extracts the corresponding ids. It skips over lines that are source() calls.
}
