% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shiny_file.R
\name{read_shiny_file}
\alias{read_shiny_file}
\title{Read a Shiny file and determine if it should be skipped}
\usage{
read_shiny_file(file_path)
}
\arguments{
\item{file_path}{A string representing the path to the R file.}
}
\value{
A character vector containing the file's lines if valid, otherwise \code{NULL}.
}
\description{
This function reads a Shiny server file and checks whether it only contains
\code{source()} calls or is empty. If so, it returns \code{NULL} to indicate the file should be skipped.
}
